Graph, Documents, Images
The graph has been simplified and now includes dates along the x axis. The graph and symptoms can be saved, send, or printed. A sqlite database has been implemented to store documents unique to each childId. Images can be taken or uploaded from the device for each child. I wasn't able to test this fully as i am using an emulator but i was able to at least get the image folder to open.